2021.10.13_박주영 fix: 당일 운동량 페이지에 물섭취량 합침 	 - asideLeft.jsp 	 - personal_management.jsp 	 - PersonalManagement.jsp 	 - sqlMapConfig.xml 	 - Water_Intake 	 - WaterIntakeController.java
